# students/KarpeshNikita/task_02/src/docs/swagger.yaml
definitions:
  models.Book:
    properties:
      author_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  models.Review:
    properties:
      book_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      rating:
        type: integer
      text:
        type: string
      user_id:
        type: integer
    type: object
  models.Shelf:
    properties:
      id:
        type: integer
      name:
        type: string
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/books:
    get:
      description: Get list of books
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List books
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Create a book (authenticated)
      parameters:
      - description: Book payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: Create a new book
      tags:
      - Books
  /api/books/{id}:
    delete:
      description: Delete book by id (authenticated)
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: Delete a book
      tags:
      - Books
    get:
      description: Get book details
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get book by id
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: Update book details (authenticated)
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Book payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: Update a book
      tags:
      - Books
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate user and return JWT
      tags:
      - Auth
  /api/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Register payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /api/reviews:
    post:
      consumes:
      - application/json
      description: Create a review for a book (authenticated)
      parameters:
      - description: Review payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: Create a review
      tags:
      - Reviews
  /api/shelves:
    get:
      description: List all shelves
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Shelf'
            type: array
      summary: List shelves
      tags:
      - Shelves
    post:
      consumes:
      - application/json
      description: Create a shelf for authenticated user
      parameters:
      - description: Shelf payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Shelf'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Shelf'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: Create a shelf
      tags:
      - Shelves
swagger: "2.0"